avaris/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
tensorflow
sqlalchemy
opencv-python
pillow
google-generativeai
python-dotenv
requests
httpx
joblib
python-multipart
